Improve support for commas in verse numbers
Commas in verse numbers are usually used to indicate verses that were
reordered into a different chapter (e.g. Verse 3:2,9 is actually 2:9, but
moved somewhere in the middle of chapter 3).

Handle this correctly in Logos milestones and versification detection.
Also add a RearrangeChapterBoundaries mode to StrippedDiffable to move
chapter boundaries in case an adjacent verse is in a different chapter
(e.g. if verse 3:2,9 is the first verse of chapter 3, make it 2:9 and move
it to end of chapter 2).