Preliminary ZefDic filter
It should be able to handle most files, some files still throw Exceptions
on import, though. I don't really find time right now to improve it (and
ZefDic is not that important anyway) so I'll stash this away in this
branch for now until I might find time to revisit it.