Add VerseRange support
a VerseRange is used to group multiple verses inside a chapter that use
overlapping verse numbers (like 1.3.5 and 2.4 gets combined to 1-5). Verse
ranges of a chapter do not overlap, but they do not need to be sorted.
This is useful for formats (like Logos) that support exactly these kind of
verse ranges.

Also add verse ranges support to versification detector framework, and
update LogosVersificationDetector and LogosHTML to use them.