Fix BrowserBible export of footnotes
If a footnote contained a paragraph break, the exported XML was invalid
(</div> did not match <span>). Reason is that paragraph breaks are
exported as closing the containing paragraph div and opening a new one,
which works fine for verses, but not for footnotes.

Fix this by exporting paragraph breaks in footnotes as two line breaks
instead.