Fix Text Comparison skipping some verses
Logos Text comparison skipped verses at the beginning of a chapter, in
case there is a headline before it and chapter headings are suppressed.

This is caused by the fact that Logos (7.12) apparently ignores the
second milestone if two milestones are next to each other or only
separated by spaces. Work around this by adding a Zero Width Non
Breaking Space control character between the two milestones.