Add LogosFoonotePostprocessor
This is a "must" if you want to view/edit the resulting .docx in any
recent MS Word version. Otherwise the footnote labels in the text and in
the footnote area on the bottom of the page do not match.

Note that this also results in the footnote labels to show up in the
tooltips within Logos, so use with care.