Add USFM import/export
This is probably the most requested feature.

Note that this importer has only been tested on very few bibles, so it
may still contain lots of bugs. Also note that some tags (\ca \cp \va
\vp \fig and \fm) are deliberately not supported (they are not supported
by any of the other formats either), and Study Bible tags are also
missing (but might be added later).

USFX and USX will come in a later commit (I have already started it but
did not want to force USFM to wait for the missing parts in these two
less requested formats).

Fixes #15.