Add USFX and USX formats
Export of USFX is not supported as the format is not clearly enough
specified, and it seems there are no programs that can read USFX but not
one of the other two Paratext formats.

Also add a ParatextDump pseudo format to compare different Paratext
format texts of the same bible without incurring the information loss
from converting to Diffable format.