Provide options to make Validate laxer
Do not validate consecutive whitespace or whitespace before/after line
breaks or other block level elements, when not desired. Or ignore empty
elements (like formattings or grammar information). That way, all the
other validations can still be performed and the whitespace errors
and/or empty elements can be cleaned up later (if wanted).