Add SoftProjector support
This format is quite simple (text only), but it supports a way to define
versification mappings so that you can view different bibles in
parallel. This is only used in very few bible files, but the concept is
interesting, so I decided to support it.

Also fix a bug: When generating verse ranges, verse numbers like 13.0
(happen sometimes from SoftProjector Modules) cause an error as they are
treated like the range 0-13, which contains invalid verses.