Change handling of Strong numbers
This is a massive commit, touching almost all format modules. Before,
the Strong numbers were treated as numbers, by default H in Old
Testament and G in New Testament, with some module specific quirks (in
case modules support to have custom prefix letters). As
@viktor-zhuromskyy pointed out, this is insufficient for modules like
LXX which have greek text in OT and which also have an extra L
dictionary in addition to G and H.

Therefore change the default that every strong reference can have a
letter A-Z attached to it and let the modules handle it. If a module
only supports G and H (or even only one kind of Strong number), the
module can ignore the prefix; other modules can use the prefix. Also
provide `StrippedDiffable` options to add prefix to Strongs (G or H
depending on Testament), remove prefixes that are default, or completely
remove Strongs numbers that have a prefix (if the export format has
trouble exporting it).

This commit probably needs some more testing before the next release :)

See #30.