Improve CSS handling in USFX export
First, add a special case to convert CSS created by MyBibleZone's `<n>`
tag to \add tag.

In addition, when there is custom CSS to be exported and there is no
exactly matching USFM tag, check if the CSS includes bold and/or italic
and fall back to \bd or \it or \bdit tags.

Finally, convert red text to \wj and small caps to \sc.

When these fuzzy rules find nothing, behave as before (print a warning
and leave the content unformatted.

See #32.