VersificationTool: Do not destroy file on error
When adding a versification to a versification database and an error
occurs (e.g. the versification already exists in the file), the tool
previously had already opened the output file, resulting in the previous
file content being lost. Now first export into a StringWriter and only
write to the file if no error occurred.