Update JAXB from 2.2.8 to 2.2.11
Yes, this is only a small step, but apparently the JAXB 2.2.8 bundled in
JDK 8 has less bugs than the official version. In particular, whitespace
nodes in mixed content are not properly preserved in JAXB 2.2.8 in the
Maven repository, but they are in 2.2.8 bundled by JDK 8.

From a quick test, 2.2.11 seems to work fine in that respect.