Improve virtual verse number printing
When a verse has the same verse number as its containing virtual verse,
its number is not printed. This causes unintended "joint verses" in case
it is not the first verse inside this virtual verse.

It was detected when importing a bible with psalm titles from
Accordance, but can happen for other import formats and scenarios as
well. In fact, it seems that I already had fixed this for a few formats
earlier.

This affects a total of 12 export formats.