Paratext: Synthesize \ft / \xt tag for footnotes
When importing to a Paratext format, footnote content is generally not
wrapped in \ft or \xt. Therefore synthesize them if the foonote does not
start with a `\f?` or `\x?` tag.

When exporting to a non-Paratext format, if the footnote content starts
with a `\ft?` or `\xt` tag, skip that tag and just export its contents.

This avoids invalid USX3 export when exporting a footnote that contains
references without the footnote text being wrapped as footnote text.