Paratext import: Fix headline moving logic
When a Paratext import is converted to a non-Paratext format, and a
verse contains a headline, immediately followed by a paragraph that
starts a new verse, the headline is moved into the verse below.

However, this logic did not work well in case the paragraph contained
part of the previous verse before starting the new verse.

Effectively,

    \p \v 1 FirstVerse
    \s1 Headline
    \p DidNotEndYet \v 2 SecondVerse

Was converted as if it was

    \p \v 1 FirstVerse
    \s1 Headline DidNotEndYet
    \p \v 2 SecondVerse

and the (mangled) headline was treated as start of verse 2.

Fixes #42.