Verbose scheme output for VersificationDetector
When passing the system property 'versificationdetector.verboseschemes'
with either 'all' or a comma separated list of schemes as value, print
more details about a versification match (of this schema).

For normal versification detector, print a list of all verses that are
missing or extra. For versification counts detector, print number of
chapters and verses for every book included.