Make sure MySword bibles contain an Abbreviation
Seems that recent versions of MySword cannot open the Bible if the
abbreviation is missing. Therefore, if the bible does not contain
abbreviation metadata, look at system property "mysword.abbreviation",
and if it is also empty, use the first letter of the bible name.

You can pass the abbreviation like this:

    java -Dmysword.abbreviation=XYZ -jar ...

See #55.