Improve BibleWorks italics support
When an italics boundary (marked by square brackets) is next to a
literal square bracket (marked by two square brackets), the output is
ambiguous and will be interpreted wrong by BibleWorks if the single
square bracket is followed by the double square bracket. Work around
this by adding a Unicode non-breaking-space between them when it
happens. Also improve the parsing so that the opposite case for the
closed brackets is parsed correctly as well.