Improve export of CSS formatting
Provide a base class `AbstractNoCSSVisitor` which will try to detect
suitable formatting instructions to replace CSS formatting. If they are
found, the corresponding `visitFormattingInstruction` is called
automatically. Regardless whether something was found or not, a new
method `visitChangedCSSFormatting` allows the base visitor to clean up
it internal state (e.g. consolidate suffix stack if a CSS formatting
resulted in multiple formatting instructions) and/or log remaining CSS.

Update all formats that can export formatting, but cannot export CSS
formatting (and that did not have a workaround for that) to use the new
base class.