Fix diffable roundtrip of empty verses
Validation will complain about empty verses, but most formats can cope
with them just fine. However, the `Diffable` format (which is often used
to store bibles with validation errors to fix them manually) could not
cope with them: When exporting, the line consisted of a verse reference
and a space (and no verse text); when trying to parse it again, the
trailing space is trimmed and the line contains not enough fields. Work
around it by explicitly adding back the space in case the field count is
2 and the original line ended with a space.