OnLineBible export: improve italics handling
Italics formatting used to be suppressed in (headlines and) footnotes.
But, in fact, italics is unsupported only in the first line of a
footnote. Therefore, only suppress it for italics tags that open on the
first line of the footnote - italics tags that open on later lines will
get converted.

Fixes #51.