Support formatted Strongs in roundtrip formats
Sometimes, roundtrip formats are not used for their format stability,
but as a convenient way to manually edit bibles. In that situation, the
fact that these formats do not support Strongs suffixes (and some of
them extract the prefixes as well) makes it harder to edit.

By setting the system property
`biblemulticonverter.roundtrip.writestrongssuffix=true`, the 6 roundtrip
formats (Compact, Diffable, RoundtripHTML, RoundtripODT,
RoundtripTaggedText, RoundtripXML) will use formatted Strongs (with
suffixes) if `biblemulticonverter.strongssuffix` is set to a supported
schema. Note that the resulting files are no longer compatible with
older versions of BibleMultiConverter (or other programs) and the XML
export does not conform to the schema. So use these options at your own
risk.

To read these files, the inverse property
`biblemulticonverter.roundtrip.parsestrongssuffix=true` must be set.