Document LibreOffice requirments for LogosHTML
LibreOffice 7.6's HTML import is broken (it seemingly randomly skips
some formatting). This was fixed in LibreOffice 24.2 (aka 7.7), but some
parts like footnote formatting changed in incompatible ways.

For now, it is advised to use LibreOffice 7.5 or older for conversion. I
will try to provide an update in the future that makes it work again
with LibreOffice 24.2.