USFX: Implement export
This export does not have as many exceptions as Haiola's USFX export
has. If you have Haiola available, prefer it to this tool when
converting from Paratext formats to USFX.

Also, when synthesizing verse end markers, avoid placing them in
introductory texts (that may appear in the middle of a chapter).