LogosHTML: Add Word 2024 mode
This mode allows using Word 2024 for conversion to .docx. This may be
more convenient for some users, and conversion is faster. Also, nested
hyperlinks do not require postprocessing. On the other hand, Logos
sometimes tends to format some whitespace as Tahoma although Times is
selected for the whole document, causing conversion warnings in Logos.

To enable, pass

    -Dbiblemulticonverter.logos.word2024=true

and convert the file using Word (tested in Word for Windows 2024). In
Word, you still need to break image links (so they get embedded) and may
need to change the font. Note that some font sizes of headlines are
different, yet the result is still usable (and you can always bring your
own stylesheets in Word).

In particular, this fixes the following issues when trying to use Word
for conversion. Some others may remain.
- Nested hyperlinks are working out of the box
- Strongs numbers on words of christ do not make the word black
- Footnotes become actual footnotes