AugmentGrammar: More dumping options
Add a mode to dump the grammar tags along with the words they are
attached to. Also add an option to dump Strongs in human format (useful
if the format uses XY Strongs).

While at it, fix converstion of RMAC from UBXF (The "robinson:" prefix
has been missing, resulting in the RMAC not being dumped).