Allow overriding Bible name
This works for all import formats. When the import format does not
provide a Bible name (or a misleading one) and the export format
requires one (which is hard to edit later), the System Property
`biblemulticonverter.bible.name` can be used to override the bible name,
therefore avoiding the need to convert to an intermediate format.

Closes #103.