Paratext export: Fix srcloc inside formatting
When a GrammarInformation appears inside another Formatting, the export
context is lost (which is correct since the export context would allow
messing with paragraphs), and the srcloc long prefix used to be part of
the export context.

Therefore, move the srcloc long prefix outside the export context and
pass it to child visitors when exporting formatting.