Export to paratext format and back lost verses
When at the beginning of a chapter, there is a headline at the beginning
of a verse, the verse marker was written at a point where the verse
marker was ignored when reading back the same file.

Fix by moving the verse marker into the paragraph after the headline
(not only in this specific case, but whenever a verse starts with a
headline).